tests-api/requirements.txt
assertpy==1.1
configparser==5.0.1
cucumber-tag-expressions==3.0.0
diffimg==0.2.3
python-dotenv==0.15.0
Faker==4.14.2
imutils==0.5.3
opencv-python==4.5.1.48
py==1.9.0
pylint==2.7.0
pytest==6.2.3
pytest-base-url==1.4.2
pytest-bdd==4.0.1
pytest-rerunfailures==9.1.1
pytest-xdist==2.1.0
pytest-logger==0.5.1
python-dateutil==2.8.1
pytest-html==3.1.1
PyYAML==5.4.0
requests==2.24.0
pytest-failed-screenshot==1.0.0
allure-pytest==2.8.40
PyHamcrest==2.0.2
jsonpath-ng==1.5.3
